requirements.txt
asgiref==3.11.0
gunicorn==23.0.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
Django==4.2.27
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-yasg==1.21.11
idna==3.11
inflection==0.5.1
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.7.0
sqlparse==0.5.4
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
